useEnsemble hook refactor
