+++ windwatts-ui/src/hooks/index.ts
@@ -4,7 +4,7 @@ export { useLocalStorage } from "./useLocalStorage";
 export { useMapView } from "./useMapView";
 export { useOutOfBounds } from "./useOutOfBounds";
 export { useProductionData } from "./useProductionData";
-export { useEnsembleTilesData } from "./useEnsembleTilesData";
+export { useEnsemble } from "./useEnsemble";
 export { useToggle } from "./useToggle";
 export { useWindData } from "./useWindData";
 export { useWindRoseData } from "./useWindRoseData";
